feat:: pixel picking

+++ week_05+/Engine/Source/Renderer/RenderCommand.cpp
@@ -4,9 +4,7 @@
 
 uint64 FRenderCommand::MakeSortKey(const FMaterial* InMaterial, const FMeshData* InMeshData, uint32 InFirstIndex)
 {
-	uint32 MatId = InMaterial ? InMaterial->GetSortId() : 0;
+	uint32 MatId = InMaterial ? static_cast<uint32>(InMaterial->GetSortId()) : 0;
 	uint32 MeshId = InMeshData ? InMeshData->GetSortId() : 0;
-
-	// MatId(32비트) | MeshId(16비트) | FirstIndex(16비트) 형태로 조합하여 충돌 방지
 	return (static_cast<uint64>(MatId) << 32) | (static_cast<uint64>(MeshId) << 16) | (InFirstIndex & 0xFFFF);
 }
